+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - The notify command now accepts a `players` argument instead of `player`
 - Updated the mod logo
+- Spirits now drop 1-2 Ectoplasm
+- Spirits are now part of the Creature MobCategory instead of MISC
 
 ### Fixed